requirements.txt
# im cooked
langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
openai==2.6.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-core==2.27.2